+++ src/test/java/TestsSportLogBook/TestsGetLogBookId.java
@@ -9,12 +9,12 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TestsGetLogBookId {
 
-    private static final int LOGBOOK_ID = 1;
+    private static final Long LOGBOOK_ID = 1L;
 
     //Проверяет возвращаемое значение getLogBookId
     @Test
     public void testGetLogBookId() {
-        SportLogBook logBook = new SportLogBook(LOGBOOK_ID, 2, 30, LocalDateTime.now(), "Running", "Good run");
+        SportLogBook logBook = new SportLogBook(LOGBOOK_ID, 2L, 30, LocalDateTime.now(), "Running", "Good run");
         assertEquals(LOGBOOK_ID, logBook.getLogBookId(), "getLogBookId вернул неверный результат");
     }
